ml/segmentation/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ml/segmentation/requirements.in
#
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.58.3
    # via matplotlib
joblib==1.5.1
    # via
    #   -r ml/segmentation/requirements.in
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.3
    # via -r ml/segmentation/requirements.in
numpy==2.3.0
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
packaging==25.0
    # via matplotlib
pandas==2.3.0
    # via -r ml/segmentation/requirements.in
pillow==11.2.1
    # via matplotlib
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via -r ml/segmentation/requirements.in
pytz==2025.2
    # via pandas
scikit-learn==1.7.0
    # via -r ml/segmentation/requirements.in
scipy==1.15.3
    # via scikit-learn
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
tzdata==2025.2
    # via pandas
